Refractoring movie container
