+++ src/componenets/App/App.js
@@ -6,6 +6,8 @@ import MovieContainer from '../../containers/MovieContainer/MovieContainer.js'
 import Login from '../Login/Login'
 import MovieDetail from '../MovieDetail/MovieDetail.js';
 import Header from '../Header/Header.js';
+import { fetchMoviesAPI, fetchRatingsAPI } from '../../apiCalls/apiCalls.js';
+import { loadingMovies, getMovies, getRatings } from '../../actions';
 
 
 class App extends Component {
@@ -14,14 +16,36 @@ class App extends Component {
     this.state = {};
   }
 
+  componentDidMount() {
+    Promise.all([this.loadMovies(), this.loadRatings()])
+  }
+
+  loadMovies() {
+    fetchMoviesAPI()
+    .then(data => {
+      this.props.getMovies(data.movies)
+      this.props.loadingMovies(true)
+    })
+  }
+
+  loadRatings() {
+    fetchRatingsAPI()
+    .then(data => {
+      this.props.getRatings(data.ratings)
+    })
+  }
+
   render() {
     return (
       <main>
         <Header />
             <div>
               <Switch>
+                <Route path='/movies/:id'
+                       render={({ match })=>< MovieDetail selectedMovie=
+                       {this.props.movies.find(movie => movie.id === parseInt(match.params.id))}/>}/>
                 <Route path='/' 
-                       render={()=><MovieContainer />} />
+                       render={()=>< MovieContainer />} />
               </Switch>
             </div>
       </main>
@@ -29,4 +53,15 @@ class App extends Component {
   }
 }
 
-export default App;
+export const mapStateToProps = (state) => ({
+  movies: state.movies 
+})
+
+const mapDispatchToProps = dispatch => ({
+  loadingMovies: (loadingStatus) => dispatch(loadingMovies(loadingStatus)),
+  getMovies: (movies) => dispatch(getMovies(movies)),
+  getRatings: (ratings) => dispatch(getRatings(ratings))
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(App);
+
