+++ src/containers/MovieContainer/MovieContainer.js
@@ -8,25 +8,7 @@ import { fetchMoviesAPI, fetchRatingsAPI } from '../../apiCalls/apiCalls.js';
 import { Loading } from '../../componenets/Loading/Loading.js';
 
 class MovieContainer extends Component {
-  componentDidMount() {
-    Promise.all([this.loadMovies(), this.loadRatings()])
-  }
-
-  loadMovies() {
-    fetchMoviesAPI()
-    .then(data => {
-      this.props.getMovies(data.movies)
-      this.props.loadingMovies(true)
-    })
-  }
-
-  loadRatings() {
-    fetchRatingsAPI()
-    .then(data => {
-      this.props.getRatings(data.ratings)
-    })
-  }
-
+  
   render() {
     let movieCards = this.props.movies.map(movie => {
       return <MovieCard
@@ -56,10 +38,10 @@ const mapStateToProps = state => {
   }
 }
 
-const mapDispatchToProps = dispatch => ({
-  loadingMovies: (loadingStatus) => dispatch(loadingMovies(loadingStatus)),
-  getMovies: (movies) => dispatch(getMovies(movies)),
-  getRatings: (ratings) => dispatch(getRatings(ratings))
-})
+// const mapDispatchToProps = dispatch => ({
+//   loadingMovies: (loadingStatus) => dispatch(loadingMovies(loadingStatus)),
+//   getMovies: (movies) => dispatch(getMovies(movies)),
+//   getRatings: (ratings) => dispatch(getRatings(ratings))
+// })
 
-export default connect(mapStateToProps, mapDispatchToProps)(MovieContainer);
+export default connect(mapStateToProps)(MovieContainer);
